requirements.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
execnet==2.1.1
executing==1.2.0
filelock==3.14.0
Flask==2.2.3
folium==0.14.0
fonttools==4.39.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
mysqlclient==2.2.4
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2024.0.1
platformdirs==4.2.2
plotly==5.13.1
pluggy==1.4.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pytest==8.1.1
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2022.7.1
pywin32==305
pyzmq==25.0.2
requests==2.28.2
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.5.0
stack-data==0.6.2
tenacity==8.2.2
tornado==6.2
traitlets==5.9.0
twitter==1.19.6
tzdata==2024.1
urllib3==1.26.15
virtualenv==20.26.2
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
XlsxWriter==3.1.0

asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
>>>>>>> 521615e (updated settings and environment variables)
